+++ src/types.jl
@@ -1,13 +1,18 @@
-const ArraySizeLike{N} = NTuple{N,Integer}
-const ArraySize{N} = NTuple{N,Int} # FIXME: Dims{N}
+# Possible argument types for specifying a piece of array shape.
+# This is the same as `eltype(ArrayShape)`.
+const ArrayShapeArg = Union{Integer,AbstractUnitRange{<:Integer}}
+
+# Possible arguments for specifying array shape.
+const ArrayShape{N} = NTuple{N,ArrayShapeArg}
+
+# Types understood as an array size. It is a rectriction of `ArrayShape{N}`.
+const ArraySizeLike{N} = NTuple{N,Union{Integer,Base.OneTo}}
 
 const ArrayAxisLike = AbstractUnitRange{<:Integer}
 const ArrayAxis = AbstractUnitRange{Int}
 
 const ArrayAxesLike{N} = NTuple{N,ArrayAxisLike}
 const ArrayAxes{N} = NTuple{N,ArrayAxis}
-
-const ArraySizeOrAxesLike{N} = Union{ArraySizeLike{N},ArrayAxesLike{N}}
 
 """
     ImageProcessing.unspecified